Methods to print out parts of the UML diagram
These methods can accurately identify every data type, access modifier and return type of fields and methods in a class, except for types that have proceeding dots, eg java.lang.String(In this case the access modifier and the return type would still be accurate but the name of the method/field would be 'lang').

Currently looking for better ways to solve this problem. Also working on displaying an actual diagram instead of using the System.out.print commands.